% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_bind.R
\name{b_bind}
\alias{b_bind}
\title{Bastiat, rbind two data frames.}
\usage{
b_bind(x, y)
}
\value{
a data frame.
}
\description{
Shortcut to rbind two data frames with possibly diverging column names. Divergent columns are filled with NA's in the frame where it is absent.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
