% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_bing_it.R
\name{b_bing_it}
\alias{b_bing_it}
\title{Bastiat, please translate using bing.}
\usage{
b_bing_it(text, origin, target)
}
\value{
Use MSFT's translation API.
}
\description{
Bastiat, please translate using bing.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
