% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_case_squad.R
\name{b_case_squad}
\alias{b_case_squad}
\title{Bastiat, please send a squad to check these cases.}
\usage{
b_case_squad(value.capture, value.reduction)
}
\value{
Have the Bastiat Detective's School latest talent investigate your case.
}
\description{
Bastiat, please send a squad to check these cases.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
