% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_clean.R
\name{b_clean}
\alias{b_clean}
\title{Bastiat, clean my string.}
\usage{
b_clean(variable = NULL, command = NULL, parameter = NULL)
}
\arguments{
\item{data.path}{Specifies where the GTA data file is located (Default: 'data/master_plus.Rdata'). Set to 'online' to download the latest copy.}
}
\value{
Output is two data frames. First data frame includes the share of each year that in intervention was in force. Second data frame states parameter choices.
}
\description{
Performs some cleaning operation on a string. It's a wrapper that helps standardise code somewhat.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
