% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_create_model_variables.R
\name{b_create_model_variables}
\alias{b_create_model_variables}
\title{Bastiat, please process my text and create the model variables.}
\usage{
b_create_model_variables(bid = NULL, evaluation = NULL, text = NULL,
  train.share = 0.82, detective.number = NULL, variables = NULL,
  dtm.incl = FALSE, dtm.metric = NULL, dtm.terms = NULL,
  acting.agency = NULL, act.values = NULL, has.value = F,
  is.td = F, keywords = F)
}
\value{
A data frame with all estimation variables
}
\description{
Bastiat, please process my text and create the model variables.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
