% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_cutoff_probability.R
\name{b_cutoff_probability}
\alias{b_cutoff_probability}
\title{Bastiat, what's the minimum prediction probability for this model?}
\usage{
b_cutoff_probability(observations = NULL, predictions = NULL,
  capture.cutoff = 0.97)
}
\value{
A scalar.
}
\description{
Bastiat, what's the minimum prediction probability for this model?
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
