% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_eval.R
\name{b_eval}
\alias{b_eval}
\title{Bastiat, evaluate this pasted command.}
\usage{
b_eval(x)
}
\value{
Whatever you sought to evaluate/execute.
}
\description{
Shortcut for "eval(parse(text=paste(x)))".
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
