% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_gta_countries.R
\name{b_gta_countries}
\alias{b_gta_countries}
\title{Bastiat, please convert this into GTA countries.}
\usage{
b_gta_countries(vector = NULL, from = NULL, to = NULL)
}
\value{
Output is two data frames. First data frame includes the share of each year that in intervention was in force. Second data frame states parameter choices.
}
\description{
Bastiat, please convert this into GTA countries.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
