% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_load_site.R
\name{b_load_site}
\alias{b_load_site}
\title{Bastiat, please wait until the following XPATH appears on the site.}
\usage{
b_load_site(xpath)
}
\arguments{
\item{xpath}{Provide the XPATH for which the function should wait.}
}
\value{
Pauses the algorithm until XPATH appears.
}
\description{
Bastiat, please wait until the following XPATH appears on the site.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
