% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_model_stats.R
\name{b_model_stats}
\alias{b_model_stats}
\title{Bastiat, generate the detective stats for this model.}
\usage{
b_model_stats(
  training.df = "pred.train",
  test.df = "pred.test",
  capture.cutoff = 0.97
)
}
\value{
A data frame including the model performance metrics.
}
\description{
Bastiat, generate the detective stats for this model.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
