% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_process_keywords.R
\name{b_process_keywords}
\alias{b_process_keywords}
\title{Bastiat, please process my text to find the GTA keywords.}
\usage{
b_process_keywords(bid = NULL, text = NULL)
}
\value{
A data frame with variables for each keyword dimension.
}
\description{
Bastiat, please process my text to find the GTA keywords.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
