% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_bt_reset.R
\name{b_reset}
\alias{b_reset}
\title{Bastiat, please reset yourself.}
\usage{
b_reset()
}
\value{
Bastiat in mint condition.
}
\description{
Takes out the garbage, un & reloads packages
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
