% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_score_me.R
\name{b_score_me}
\alias{b_score_me}
\title{Bastiat, please score this detective's performance.}
\usage{
b_score_me(value.capture, value.reduction)
}
\value{
Get a value between 0 and 1 for how good he did.
}
\description{
Bastiat, please score this detective's performance.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
