% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_train_detective.R
\name{b_train_detective}
\alias{b_train_detective}
\title{Bastiat, please train this detective.}
\usage{
b_train_detective(detective.number = NULL, variables = NULL,
  estimation.method = NULL, estimation.rounds = NULL,
  dtm.incl = FALSE, dtm.metric = NULL, dtm.terms = NULL,
  train.share = NULL, practice = T, tested.models = c("XGB", "KNN",
  "RNF", "SVM", "SL"))
}
\value{
Several data frames about how the detective did.
}
\description{
Bastiat, please train this detective.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
