% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_update.R
\name{b_update}
\alias{b_update}
\title{Get the latest version from GitHub.}
\usage{
b_update(x)
}
\value{
Be up to date with our latest functions.
}
\description{
Syncs your local library with our latest GTA GitHub release.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
