% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_var_generation.R
\name{b_var_generation}
\alias{b_var_generation}
\title{Bastiat, draw new variables for my model.}
\usage{
b_var_generation(x, method = "fully random")
}
\value{
A vector of variable names.
}
\description{
Have Bastiat specify your learning model for you
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
