% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_var_pool.R
\name{b_var_pool}
\alias{b_var_pool}
\title{Bastiat, pool variables from two models.}
\usage{
b_var_pool(x, y, uncommon.selection = "random")
}
\value{
A vector of variable names.
}
\description{
Create a model based on two old ones (plus some randomness).
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
