% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bing_it.R
\name{bing_it}
\alias{bing_it}
\title{Bastiat, please translate using bing (legacy function).}
\usage{
bing_it(text, origin, target)
}
\value{
Use MSFT's translation API.
}
\description{
Bastiat, please translate using bing (legacy function).
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
