% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_classify_by_keyword.R
\name{bt_classify_by_keyword}
\alias{bt_classify_by_keyword}
\title{Bastiat, please estimate a new squad classifier based on the best detectives.}
\usage{
bt_classify_by_keyword(
  text.to.check = NULL,
  text.id = NULL,
  flavour = "negative"
)
}
\value{
Several data frames about how the detective did.
}
\description{
Bastiat, please estimate a new squad classifier based on the best detectives.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
