% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_collect_url.R
\name{bt_collect_url}
\alias{bt_collect_url}
\title{Bastiat, collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.}
\usage{
bt_collect_url(
  url = NULL,
  file.name = NULL,
  store.path = NULL,
  add.timestamp = F,
  update.file.name = F,
  js.path = "setup/rasterize.js",
  phantom.path = "~/bin/phantomjs"
)
}
\value{
#returns a list of 4 objs: new.file.name, file.suffix, url, status {see gta_prod.gta_url_status_list for these}
}
\description{
Bastiat, collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz, Robin Scherrer and Callum Campbell for GTA
}
