% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_create_estimation_data.R
\name{bt_create_estimation_data}
\alias{bt_create_estimation_data}
\title{Bastiat, please process my text and create the model variables.}
\usage{
bt_create_estimation_data(
  bid = NULL,
  evaluation = NULL,
  word.score = NULL,
  text = NULL,
  acting.agency = NULL,
  train.share = 0.82,
  detective.name = NULL,
  detective.number = NULL,
  for.training = T,
  max.input.rows = 20000
)
}
\value{
A list including data frames for the estimation data, the word scores and a vector of variables.
}
\description{
Bastiat, please process my text and create the model variables.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
