% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_detective_prediction.R
\name{bt_detective_prediction}
\alias{bt_detective_prediction}
\title{Bastiat, have the detective predice this data.}
\usage{
bt_detective_prediction(
  detective = NULL,
  classifier.location = NULL,
  prediction.data.id = NULL,
  prediction.word.score = NULL,
  prediction.data.text = NULL,
  prediction.acting.agency = NULL,
  train.share = 0.82
)
}
\value{
A data frame with the predictions.
}
\description{
Bastiat, have the detective predice this data.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
