% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_encoding_brute_force.R
\name{bt_encoding_brute_force}
\alias{bt_encoding_brute_force}
\title{Try conversion to every possible encoding}
\usage{
bt_encoding_brute_force(tgt.string, print.all = F)
}
\arguments{
\item{tgt.string}{the string to convert}

\item{print.all}{print all the results. a human will probably have to read
this to determine the answer.}
}
\value{
a list of all the results
}
\description{
Tries converting the target string from every single encoding into UTF8.
}
\details{
Determines once and for all if conversion is the problem, or the original is
simply corrupted.

Probably requires human looking at the results to see which (if any) of the
encodings produced a salient result.
}
\references{
www.globaltradealert.org
}
