% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_estimate_classifier.R
\name{bt_estimate_classifier}
\alias{bt_estimate_classifier}
\title{Bastiat, please estimate this classifier.}
\usage{
bt_estimate_classifier(
  training.data = NULL,
  training.id = "bid",
  training.y = "evaluation",
  estimation.model = "XGB",
  robustness.turns = 0,
  train.share = 0.82,
  update.log = F,
  detective.name = NULL,
  detective.number = NULL,
  max.training.rows = 15000
)
}
\value{
A list including the classifier object plus a performance data frame.
}
\description{
Bastiat, please estimate this classifier.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA.
}
