% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_estimate_detective.R
\name{bt_estimate_detective}
\alias{bt_estimate_detective}
\title{Bastiat, please re-estimate this detective.}
\usage{
bt_estimate_detective(
  detective.name = NULL,
  robustness.turns = 0,
  train.share = 0.82,
  save.stats = T,
  save.classifier = T,
  training.data.path = "data/classifier/training data.Rdata",
  training.data.update = F
)
}
\value{
Several data frames about how the detective did.
}
\description{
Bastiat, please re-estimate this detective.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
