% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_estimate_news_leads.R
\name{bt_estimate_news_leads}
\alias{bt_estimate_news_leads}
\title{Generates predictions for relevance using the most recent Mrs Hudson model.}
\usage{
bt_estimate_news_leads(
  leads.core.news,
  keep.results.ratio = 0.95,
  binary.prediction = T,
  return.both = F,
  conf.cutoff = 0.2
)
}
\arguments{
\item{leads.core.news}{A dataframe of leads in the standard format. Requires
acting.agency, act.title.en, act.description.en}

\item{keep.results.ratio}{what fraction of the results you want to keep. 1 =
keep all, 0 = discard all. the ones with lowest relevance prob will be
discarded.}

\item{binary.prediction}{return results as a binary prediction or as their
raw 0 ~ 1 confidence values.}

\item{return.both}{return both binary and raw.score in a list.}
}
\value{
A vector of Mrs. Hudson's predictions.
}
\description{
Generates predictions for relevance using the most recent Mrs Hudson model.
}
\references{
www.globaltradealert.org
}
