% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_estimate_squad.R
\name{bt_estimate_squad}
\alias{bt_estimate_squad}
\title{Perform estimation of squad classifier}
\usage{
bt_estimate_squad(
  detectives = NULL,
  squad.level = 0.7,
  train.share = 0.82,
  squad.model = "XGB",
  squad.robustness = 0,
  detective.log = "content/0 core/Classifier statistics & history.Rdata",
  squad.log = "content/0 core/Squad statistics & history.Rdata",
  estimation.data = "data/classifier/training data.Rdata",
  limit.classification = 50000
)
}
\arguments{
\item{detectives}{the squad to be estimated}

\item{squad.level}{min threshold for the squad score}

\item{train.share}{train:test split ratio}

\item{squad.model}{which model to use for the squad}

\item{detective.log}{location of the log file}

\item{squad.log}{location of the squad log file}

\item{estimation.data}{the data to be used for testing}

\item{limit.classification}{reduces the amount of rows of data to test to reduce comp load (my 4GB ram laptop tends to error above ~50k rows hence this being the default value)}
}
\value{
Several data frames about how the detective did.

Several data frames about how the detective did.
}
\description{
Perform estimation of squad classifier
}
\references{
www.globaltradealert.org

www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
Bastiat, please estimate a new squad classifier based on the best detectives.

Johannes Fritz for GTA
}
