% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_extract_url.R
\name{bt_extract_url}
\alias{bt_extract_url}
\title{Bastiat, please extract all URLs you find in the following string.}
\usage{
bt_extract_url(string, get.tld = F)
}
\description{
Bastiat, please extract all URLs you find in the following string.
}
