% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_gen_d2v_conc.R
\name{bt_gen_d2v_conc}
\alias{bt_gen_d2v_conc}
\title{Generate concatenated matrix of doc2vec vectors for Mrs Hudson}
\usage{
bt_gen_d2v_conc(leads.core)
}
\arguments{
\item{leads.core}{}
}
\value{

}
\description{
Vectorises text using saved training colnames. preprocesses the text as well.
}
