% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_generate_2cat_tf_idf.R
\name{bt_generate_2cat_tf_idf}
\alias{bt_generate_2cat_tf_idf}
\title{Get tf-idf scores for terms two categories of document and create master list}
\usage{
bt_generate_2cat_tf_idf(unnested.token.df, rlv.doc.ids)
}
\arguments{
\item{unnested.token.df}{product of bt_unnest_tokens()}

\item{rlv.doc.ids}{the ids of the relevant (category 1) documents (others will be irrelevant/cat 2)}
}
\description{
Get tf-idf scores for terms two categories of document and create master list
}
