% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_generate_tf_idf_agg_score.R
\name{bt_generate_tf_idf_agg_score}
\alias{bt_generate_tf_idf_agg_score}
\title{Get the tf-idf-type scores per document using master vocab list}
\usage{
bt_generate_tf_idf_agg_score(tf.idf.master.path, doc.id, text)
}
\arguments{
\item{text}{}
}
\description{
Get the tf-idf-type scores per document using master vocab list
}
\examples{
bt_generate_tf_idf_agg_score(tf.idf.file.name,
doc.id = leads.core.b221$bid,
text = paste(bt_text_preprocess(leads.core.b221$hint.title),
             bt_text_preprocess(leads.core.b221$hint.description)))
}
