% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_get_detective_characteristics.R
\name{bt_get_detective_characteristics}
\alias{bt_get_detective_characteristics}
\title{Bastiat, please retrieve the characteristics of this detective.}
\usage{
bt_get_detective_characteristics(
  d.name = NULL,
  d.number = NULL,
  detective.log = "content/0 core/Classifier statistics & history.Rdata"
)
}
\value{
A list of characteristics.
}
\description{
Bastiat, please retrieve the characteristics of this detective.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
