% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_get_max_bid.R
\name{bt_get_max_bid}
\alias{bt_get_max_bid}
\title{Get the maximum numbered entry for a given BID stem}
\usage{
bt_get_max_bid(bid.stem, force.con.open = F)
}
\arguments{
\item{bid.stem}{the stem to search for in the db}

\item{force.con.open}{open the db connection in the function call}
}
\description{
Get the maximum numbered entry for a given BID stem
}
