% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_get_squad_cutoff.R
\name{bt_get_squad_cutoff}
\alias{bt_get_squad_cutoff}
\title{Returns current cutoff value used by the BT squad}
\usage{
bt_get_squad_cutoff(
  squad.classifier = "content/0 core/Bastiat squad classifier.Rdata"
)
}
\value{
the cutoff value used by the squad to determine relevance
}
\description{
Like a standard getter function.
}
