% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_guess_date.R
\name{bt_guess_date}
\alias{bt_guess_date}
\title{Guesses the date from a string}
\usage{
bt_guess_date(tgt.string)
}
\arguments{
\item{tgt.string}{a string containing a potential date in the appropriate
format}
}
\value{
date in date format
}
\description{
Feed me a string with a date and I'll try to return a date formatted value.
}
\details{
The target string should contain a day value as a 1- or 2-digit number, a
year as a 4-digit number and a month as a word (e.g. 'December').

The month guessing uses bt_guess_month() and thus supports all the languages
which that function can use.

I have not seen any instances where the date is stored with the year as two
digits. If this is problematic I can implement something to catch these
cases.
}
\examples{
bt_guess_month("23eme Aout, 2021")
#returns "2021-08-23"
}
