% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_kill_runaway_pjs.R
\name{bt_kill_runaway_pjs}
\alias{bt_kill_runaway_pjs}
\title{Tries to kill extraneous PJS processes on UNIX}
\usage{
bt_kill_runaway_pjs()
}
\description{
Sometimes, pjs processes are left running after crashing or not being
terminated properly and can eat up the CPU. This function tries to kill them
to free up memory.
}
