% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_leads_classify_only.R
\name{bt_leads_classify_only}
\alias{bt_leads_classify_only}
\title{The classification part of bt_leads_core_update()}
\usage{
bt_leads_classify_only(update.core)
}
\arguments{
\item{update.core}{standard update.core with the standard variables.}
}
\value{
a new update.core with probabilities
}
\description{
Designed for use on DPA leads. Gives them a relevance probability using BT's standard squad prediction.
The code is taken directly from bt_leads_core_update to preserve similarity.
}
\examples{
update.core = bt_leads_classify_only(update.core)
}
