% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_leads_core_update.R
\name{bt_leads_core_update}
\alias{bt_leads_core_update}
\title{Get the latest version from GitHub.}
\usage{
bt_leads_core_update(
  update.df = NULL,
  exclude.by = "act.url",
  all.covid = F,
  force.create = F,
  set.official = T,
  destination = "b221",
  incl.kanji = F,
  invoke.mrs.hudson = T,
  mrs.hudson.keep.results.ratio = 0.95
)
}
\value{
Be up to date with our latest functions.
}
\description{
Syncs your local library with our latest GTA GitHub release.
}
