% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_leads_df_kw_filter.R
\name{bt_leads_df_kw_filter}
\alias{bt_leads_df_kw_filter}
\title{Keep/discard leads containing specified keywords
Use bt_update_core_kw_filter() for setting the relevant/classify instead}
\usage{
bt_leads_df_kw_filter(
  leads.df,
  key.words,
  positive = T,
  filter.cols = c("act.title.en", "act.description.en"),
  case.sensitive = F
)
}
\arguments{
\item{leads.df}{dataframe to be filtered}

\item{key.words}{character vector of words/phrases to filter (regex)}

\item{positive}{if F, entries containing the key.words will be discarded.}

\item{filter.cols}{which columns to use for filtering}
}
\value{
filtered dataframe
}
\description{
Keep/discard leads containing specified keywords
Use bt_update_core_kw_filter() for setting the relevant/classify instead
}
\examples{
bt_leads_df_kw_filter(update.table, c("(import|export) tariff", "illegal trade practices"), filter.cols = "act.title.en")
}
