% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_load_site.R
\name{bt_load_site}
\alias{bt_load_site}
\title{An updated version of b_load_site}
\usage{
bt_load_site(
  xpath = NULL,
  wait = 30,
  abort = F,
  wait.interval = 1.5,
  assign.html = F
)
}
\arguments{
\item{xpath}{Provide the XPATH for which the function should wait.}

\item{wait}{Specify the number of seconds you want to wait. Default is 60.}

\item{abort}{Specifiy whether to cause an error in case path is not find
after waiting time. Default is FALSE.}
}
\value{
Pauses the algorithm until XPATH appears.
}
\description{
Waits for the specified xpath node to be present in the HTML of webdriver's
current page.
}
\details{
Requires webdriver to be in use, assigned to an object called 'remDr'.
}
\references{
www.globaltradealert.org
}
