% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_sa_record_new_source.R
\name{bt_sa_record_new_source}
\alias{bt_sa_record_new_source}
\title{Checks state acts for URLs and adds them to the tables for scraping later on.
Step 1 in source collection process.}
\usage{
bt_sa_record_new_source(
  establish.connection = T,
  timeframe = NA,
  recheck.existing.sources = F,
  ignore.manually.added = T,
  exclude.data.dump.measures = T,
  specific.sa.ids = NA
)
}
\arguments{
\item{establish.connection}{If T, opens the DB connection.}

\item{timeframe}{Search for sources in the last {timeframe} days.}

\item{recheck.existing.sources}{If T, try to scrape sources for entries which have already been added to gta_url_log.}

\item{ignore.manually.added}{If T, do not search for sources for measures that have an associated file in gta_files. Note that there is no way to distinguish if a file is present for a particular SOURCE.}
}
\description{
Checks state acts for URLs and adds them to the tables for scraping later on.
Step 1 in source collection process.
}
