% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_scraper_checkpoint_0.R
\name{bt_scraper_checkpoint_0}
\alias{bt_scraper_checkpoint_0}
\title{Check a scraper has initialised}
\usage{
bt_scraper_checkpoint_0()
}
\value{
'OK!' if all is well, error if not.
}
\description{
Ensures:
pjs initialised with browser object called 'remDr'
main.url properly initialised
stock.data.file is specified
}
