% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_scraper_checkpoint_1.R
\name{bt_scraper_checkpoint_1}
\alias{bt_scraper_checkpoint_1}
\title{Check the initial scrape worked properly}
\usage{
bt_scraper_checkpoint_1()
}
\value{
'OK!' if all is well, error if not.
}
\description{
This function ensures:
getArticles() exists
table.main exists
the table.main has the 5 vars it should, correctly named and in the correct format
(this is particularly pertinent for the act.date)
}
\details{
the required vars are:
"act.date", "act.title.en", "act.description.en", "act.url", "act.id"
}
