% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_scraper_checkpoint_2.R
\name{bt_scraper_checkpoint_2}
\alias{bt_scraper_checkpoint_2}
\title{checks scraped data is ready for upload}
\usage{
bt_scraper_checkpoint_2(ignore.url = F, ignore.titles = F)
}
\arguments{
\item{ignore.url}{set this as T if you know your source only has one URL shared for all leads.}
}
\value{
'OK!' if all is well, error if not.
}
\description{
Ensures:
update.table exists
the 17 required vars are there
the date is correctly formatted as a date
the BIDs are unique
the titles are unique
the URLs are unique
}
\details{
17 required vars are:
"act.date", "act.title.en", "act.description.en","act.title.ll", "act.description.ll",
"act.url", "act.id", "country", "country.lead","email.language","act.values", "background.url",
"acting.agency", "relevant", "classify","collection.date", "bid"
}
