% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_scraper_init.R
\name{bt_scraper_init}
\alias{bt_scraper_init}
\title{Initialise the environment before scrape routine}
\usage{
bt_scraper_init()
}
\value{

}
\description{
Loads all packages and opens the DB connection in preparation for running the scrape routine.
}
\details{
Also cleans the environment - so make sure you save any objects you need before running this!

Any packages the routine requires should go in here.

This will only work in Bastiat's home directory.

Originally was located at the beginning of bt_scrape_handler.R
}
