% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_scraper_master_list.R
\name{bt_scraper_master_list}
\alias{bt_scraper_master_list}
\title{Show/update DB for the master list of GTA scrapers}
\usage{
bt_scraper_master_list(
  show.stock.data.errors = F,
  update.db = F,
  quiet = T,
  incl.under.maintenance = F
)
}
\arguments{
\item{show.stock.data.errors}{show scrapers for which stock data could not be loaded. Inaccurate as some scrapers work differently (e.g. EIB).}

\item{update.db}{send the updated list of scrapers to the database.}
}
\value{
a data.frame with the collated scraper information.
}
\description{
Analyses codes in the daily/running folder to show an overview of all GTA scrapers.
}
