% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_squad_prediction.R
\name{bt_squad_prediction}
\alias{bt_squad_prediction}
\title{Bastiat, please use the squad classifier to identify relevant text.}
\usage{
bt_squad_prediction(
  prediction.data.id = NULL,
  prediction.data.text = NULL,
  prediction.acting.agency = NULL,
  train.share = 0.82,
  squad.classifier = "content/0 core/Bastiat squad classifier.Rdata"
)
}
\value{
A data frame with the text id and the classification result.
}
\description{
Bastiat, please use the squad classifier to identify relevant text.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
