% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_store_sa_source.R
\name{bt_store_sa_source}
\alias{bt_store_sa_source}
\title{Bastiat, please extract all URLs you find in the following string.}
\usage{
bt_store_sa_source()
}
\description{
Bastiat, please extract all URLs you find in the following string.
}
