% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_sync_b221_main.R
\name{bt_sync_221_main}
\alias{bt_sync_221_main}
\title{Bastiat, please extract all URLs you find in the following string.}
\usage{
bt_sync_221_main()
}
\description{
Bastiat, please extract all URLs you find in the following string.
}
