% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_translate.R
\name{bt_translate}
\alias{bt_translate}
\title{Wrapper for google translate to prevent overtranslation}
\usage{
bt_translate(string, trunc = 275)
}
\arguments{
\item{string}{string to be translated}

\item{trunc}{amount of characters to limit it to}
}
\value{
exactly the same as googleLanguageR::gl_translate
}
\description{
after some analysis, it was determined that 200-300 is the best amount of
characters to translate to make accurate predictions, see
projects/translation stats for some graphs and stuff.
}
