% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_upload_to_aws.R
\name{bt_upload_to_aws}
\alias{bt_upload_to_aws}
\title{Bastiat, upload a file into our AWS S3 account and return its  public URL.}
\usage{
bt_upload_to_aws(
  upload.file.name = NULL,
  upload.file.path = NULL,
  upload.destination = "gtaupload/Uploads/files/",
  credential.path = "setup/keys/as3.R",
  rm.after.upload = T
)
}
\value{
An AWS url.
}
\description{
Bastiat, upload a file into our AWS S3 account and return its  public URL.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
