% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_xpath.R
\name{bt_xpath}
\alias{bt_xpath}
\title{Bastiat, extract from an XPATH but keep missing nodes as NA's.}
\usage{
bt_xpath(path.base = NULL, path.extension = NULL, get.attribute = F,
  the.attribute = "href")
}
\arguments{
\item{path.base}{Enter a portion of the path that you are sure that exists.}

\item{path.extension}{Enter the remainder where there may be gaps wrt to final node you are interested in.}

\item{get.attribute}{Do you want to extract the value from an attribute? Default is 'FALSE'.}

\item{the.attribute}{Name the attribute you want to extract the value from. Default is 'href'.}
}
\value{
A vector of values.
}
\description{
Performs some cleaning operation on a string. It's a wrapper that helps standardise code somewhat.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA based on 'hrbrmstr' in https://stackoverflow.com/questions/32429574
}
