% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_xpath_nodes_intersect.R
\name{bt_xpath_nodes_intersect}
\alias{bt_xpath_nodes_intersect}
\title{Get the intersection of two sets of xpath nodes}
\usage{
bt_xpath_nodes_intersect(ns1, ns2)
}
\arguments{
\item{ns1}{first node set}

\item{ns2}{second node set}
}
\value{
common nodes to the two sets
}
\description{
Get the intersection of two sets of xpath nodes
}
