% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_clean.R
\name{b_clean}
\alias{b_clean}
\title{Bastiat, clean my string.}
\usage{
b_clean(variable = NULL, command = NULL, parameter = NULL)
}
\arguments{
\item{variable}{Specify the vector on which to apply the cleaning operation.}

\item{command}{Specify the command you want to use for the cleaning.}

\item{parameter}{Give the parameter commands to do the cleaning.}
}
\value{
Clean text.
}
\description{
Performs some cleaning operation on a string. It's a wrapper that helps standardise code somewhat.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
