% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_clean.R
\name{b_clean}
\alias{b_clean}
\title{Bastiat, clean my string.}
\usage{
b_clean(variable = NULL, command = NULL, parameter = NULL)
}
\value{
Clean text.
}
\description{
Performs some cleaning operation on a string. It's a wrapper that helps standardise code somewhat.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
