% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_collect_url.R
\name{bt_collect_url}
\alias{bt_collect_url}
\title{Bastiat, collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.}
\usage{
bt_collect_url(
  url = NULL,
  file.name = NULL,
  store.path = NULL,
  add.timestamp = T,
  update.file.name = T,
  js.path = "setup/rasterize.js",
  phantom.path = "~/bin/phantomjs"
)
}
\value{
The complete file name of the collected url (incl. file format).
}
\description{
Bastiat, collects a URL and stores it locally using file name. Websites are printed into PDF, URLS to files are downloaded.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
