% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_estimate_squad.R
\name{bt_estimate_squad}
\alias{bt_estimate_squad}
\title{Bastiat, please estimate a new squad classifier based on the best detectives.}
\usage{
bt_estimate_squad(
  detectives = NULL,
  squad.level = 0.7,
  train.share = 0.82,
  squad.model = "XGB",
  squad.robustness = 0,
  detective.log = "content/0 core/Classifier statistics & history.Rdata",
  squad.log = "content/0 core/Squad statistics & history.Rdata",
  estimation.data = "data/classifier/training data.Rdata"
)
}
\value{
Several data frames about how the detective did.
}
\description{
Bastiat, please estimate a new squad classifier based on the best detectives.
}
\references{
www.globaltradealert.org
}
\author{
Johannes Fritz for GTA
}
