% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_guess_month.R
\name{bt_guess_month}
\alias{bt_guess_month}
\title{Guesses month index of several languages.}
\usage{
bt_guess_month(tgt.string)
}
\arguments{
\item{tgt.string}{string containing month in supported languages}
}
\value{
integer number of guessed month
}
\description{
Works on string or char vector.
Currently German, French, Spanish, Russian, Bahasa, Italian, Arabic-gregorian, Dutch
}
\details{
Only first result returned if there are multiple matches.

NB CJK languages dates are 1月，2月，3月。。。12月
so can use str_extract(string, "\\d{1,2}") or equivalent
}
\references{
www.globaltradealert.org
}
