% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_leads_core_update.R
\name{bt_leads_core_update}
\alias{bt_leads_core_update}
\title{Send a prepared leads.core dataframe to the ricardo database
see parameter explanations for some help. bt_leads_classify_only() also available if you only want the rel probs.}
\usage{
bt_leads_core_update(
  update.df = NULL,
  exclude.by = "act.url",
  all.covid = F,
  force.create = F,
  set.official = F,
  destination = "b221",
  incl.kanji = F,
  invoke.mrs.hudson = T,
  mrs.hudson.keep.results.ratio = 1,
  classify.master.on = F
)
}
\arguments{
\item{update.df}{the data frame containing leads to update. min 17 specifically named cols}

\item{exclude.by}{unique identifier for each row; 'primary key'}

\item{all.covid}{are these all to do with you know what}

\item{set.official}{is the officiality of the source designated in the df?}

\item{destination}{"b221" -> ricardo; "leads" -> main site; "parking" -> state 10; "5" -> state 5; "dpa" -> hint_type_id = 3 = dpa}

\item{incl.kanji}{does the source include CJK chars? (kanji, hanzi, kana, hangeul)}

\item{invoke.mrs.hudson}{use to disable mrs. h invocation}

\item{mrs.hudson.keep.results.ratio}{the ratio of results for Mrs. H to keep. eg. 0.9 keeps 90% of them.}

\item{classify.master.on}{if T, runs the R-based classification system. If F, puts rel_prob as NA for the AWS stack to do it.}
}
\description{
Send a prepared leads.core dataframe to the ricardo database
see parameter explanations for some help. bt_leads_classify_only() also available if you only want the rel probs.
}
\examples{
bt_leads_core_update("update.core", "bid", destination = "b221")
}
