% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_text_preprocess.R
\name{bt_text_preprocess}
\alias{bt_text_preprocess}
\title{Basic preprocessing of text for classification/training}
\usage{
bt_text_preprocess(text, stop.rm = T)
}
\arguments{
\item{text}{string or character vector}

\item{stop.rm}{remove stopwords. set to F for e.g. acting.agency}
}
\value{

}
\description{
Takes text and returns it lowercase, minus punctuation and stopwords.
}
