% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_update_core_kw_filter.R
\name{bt_update_core_kw_filter}
\alias{bt_update_core_kw_filter}
\title{Set relevant/classify based on keywords
Use bt_leads_df_kw_filter() keeping/discarding leads based on keywords}
\usage{
bt_update_core_kw_filter(
  update.core,
  key.words,
  positive = T,
  filter.cols = c("act.title.en", "act.description.en", "act.title.ll",
    "act.description.ll"),
  case.sensitive = F,
  only.keep.matches = T,
  save.match.words = F,
  assign.relevance = T
)
}
\arguments{
\item{update.core}{dataframe to be filtered}

\item{key.words}{character vector of words/phrases to filter (regex)}

\item{positive}{if F, entries containing the key.words will be discarded.}

\item{filter.cols}{which columns to use for filtering}

\item{only.keep.matches}{if True, sets all relevant/classify to 0 first}

\item{save.match.words}{if True, add a column "key.words" containing all the words that generated matches}

\item{assign.relevance}{if True, assign the relevant and classify 0/1 to the returned df. default is FALSE}
}
\value{
filtered dataframe
}
\description{
Set relevant/classify based on keywords
Use bt_leads_df_kw_filter() keeping/discarding leads based on keywords
}
\examples{
bt_update_core_kw_filter(update.core, c("(import|export) tariff", "illegal trade practices"), filter.cols = "act.title.en")
}
